feat: Restructure project into a Python package, implement new NSE client, utility functions, and agent architecture, and update Docker configurations.
