+++ .env.sample
@@ -1,6 +1,17 @@
+# Required for ADK
+LITE_LLM_MODEL=provider/model-name
+LITE_LLM_API_KEY=X
+
+# Optional: Required for Google Services overrides
 GOOGLE_API_KEY=X
+
+# Optional: Required only for Reddit Sentiment features
 REDDIT_CLIENT_ID=X
 REDDIT_CLIENT_SECRET=X
 REDDIT_USER_AGENT=stock-analysis-mcp/1.0
-LITE_LLM_MODEL=provider/model-name
-LITE_LLM_API_KEY=X
+
+# Optional Configuration Defaults
+HOST=127.0.0.1
+PORT=8080
+MCP_URL=http://mcp:8000/sse
+LOG_FORMAT=color
