--- Dockerfile
@@ -1,10 +1,0 @@
-FROM python:3.11-slim
-
-WORKDIR /app
-
-COPY requirements.txt requirements.txt
-RUN pip install --no-cache-dir -r requirements.txt
-
-COPY . .
-
-CMD python server.py & adk web --host 0.0.0.0 --port 8080
