--- docker-compose.yml
@@ -1,20 +1,0 @@
-version: '3.8'
-
-services:
-  nseindia:
-    image: imcodeman/nseindia
-    ports:
-      - "3001:3001"
-
-  mcp:
-    build: .
-    environment:
-      - GOOGLE_API_KEY=$GOOGLE_API_KEY
-      - REDDIT_CLIENT_ID=$REDDIT_CLIENT_ID
-      - REDDIT_CLIENT_SECRET=$REDDIT_CLIENT_SECRET
-      - REDDIT_USER_AGENT=$REDDIT_USER_AGENT
-    ports:
-      - "8000:8000"
-      - "8080:8080"
-    depends_on:
-      - nseindia
