+++ stock_analysis_mcp/adk_service.py
