+++ stock_analysis_mcp/agent/tools.py
@@ -12,8 +12,8 @@ import matplotlib.pyplot as plt
 from google.adk.tools import ToolContext
 from google.adk.tools.agent_tool import AgentTool
 
-from agent.subagent import generate_plot_code_agent
-from logging_config import setup_logging
+from stock_analysis_mcp.agent.subagent import generate_plot_code_agent
+from stock_analysis_mcp.logging_config import setup_logging
 
 setup_logging()
 logger = logging.getLogger("tools")
