+++ stock_analysis_mcp/constants.py
