+++ stock_analysis_mcp/server.py
@@ -9,7 +9,7 @@ from fastmcp import FastMCP
 from starlette.requests import Request
 from starlette.responses import JSONResponse, Response
 
-from utils import (
+from stock_analysis_mcp.utils import (
     get_chaikin_money_flow,
     get_data,
     get_ema,
